+++ include/dropout.hpp
@@ -48,12 +48,12 @@ public:
         Matrix dropped(input.size(), std::vector<Weight>(input[0].size(), 0));
         Matrix mask(input.size(), std::vector<Weight>(input[0].size(), 0));
         // generate mask of bernoulli variables. Scale by the keep_prob.
-        for(int row = 0; row < mask.size(); ++row) {
+        for(size_t row = 0; row < mask.size(); ++row) {
             std::fill(mask[row].begin(), mask[row].end(), 
                       static_cast<Weight>(keep()) / keep_prob);
         }
-        for(int row = 0; row < input.size(); ++row) {
-            for(int col = 0; col < input[0].size(); ++col) {
+        for(size_t row = 0; row < input.size(); ++row) {
+            for(size_t col = 0; col < input[0].size(); ++col) {
                 dropped[row][col] = input[row][col] * mask[row][col];
             }
         }
@@ -69,8 +69,8 @@ public:
 
     Matrix backward_pass(const Matrix& d_out) {
         Matrix d_input(d_out.size(), std::vector<Weight>(d_out[0].size(), 0)); 
-        for(int i = 0; i < d_out.size(); ++i) {
-            for(int j = 0; j < d_out[0].size(); ++j) {
+        for(size_t i = 0; i < d_out.size(); ++i) {
+            for(size_t j = 0; j < d_out[0].size(); ++j) {
                 d_input[i][j] = d_out[i][j] * this->last_output[i][j];
             }
         }
